requirements.txt
ntplib==0.3.3
numpy==1.16.1
pyparsing==2.3.1
python-dateutil==2.8.0
pyzmq==17.1.2
six==1.12.0
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
